Inline cumulus geotag controller in top-level geotag controller
