--- collections-online-cumulus/plugins/geo-tag-controller.js
@@ -1,7 +1,0 @@
-'use strict';
-
-module.exports = {
-  type: 'geo-tag-controller',
-  module: require('../controllers/geo-tagging'),
-  initialize: () => {}
-};
